.github/scripts/requirements.txt
rich
requests
typer
PyYAML
unidecode
questionary
biopython